+++ tests/test_builder.py
@@ -13,7 +13,7 @@ from justhtml.node import Comment, Document, DocumentFragment, Element, Template
 
 
 def JustHTML(*args, **kwargs):  # noqa: N802
-    if "sanitize" not in kwargs and "safe" not in kwargs:
+    if "sanitize" not in kwargs:
         kwargs["sanitize"] = False
     return _JustHTML(*args, **kwargs)